refactor: move interfaces to separate file
