+++ src/components/main/main-content.css
@@ -10,7 +10,7 @@
 }
 
 .card {
-  width: 350px;
+  width: 400px;
   padding: 50px;
 
   color: aliceblue;
@@ -18,4 +18,5 @@
 
   border: 2px lightsteelblue solid;
   border-radius: 20px;
+  background-color: rgba(0, 0, 0, 0.5);
 }
